feat: Divide SearchBar component into Main and Portfolio
