+++ src/pages/MainPage.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import GNBBOX from "../components/common/GNBBOX";
 import MainCarousel from "../components/main/MainCarousel";
 import MainHeaderRow from "../components/main/MainHeaderRow";
-import SearchBar from "../components/common/SearchBar";
+import MainSearchBar from "../components/main/MainSearchBar";
 
 export default function MainPage() {
   const [isSearchBarOpen, setIsSearchBarOpen] = useState(false);
@@ -18,7 +18,7 @@ export default function MainPage() {
     <div className="flex w-full h-full flex-col">
       <div className="w-full h-full ">
         {isSearchBarOpen && (
-          <SearchBar handleCloseSearchBar={handleCloseSearchBar} />
+          <MainSearchBar handleCloseSearchBar={handleCloseSearchBar} />
         )}
         {!isSearchBarOpen && (
           <MainHeaderRow handleOpenSearchBar={handleOpenSearchBar} />
